requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
discord.py==1.2.5
idna==2.8
multidict==4.7.2
pathspec==0.6.0
praw==6.4.0
prawcore==1.0.1
regex==2019.12.20
requests==2.22.0
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
update-checker==0.16
urllib3==1.25.7
websocket-client==0.57.0
websockets==6.0
yarl==1.4.2
